+++ CHANGES.md
@@ -2,6 +2,7 @@ Change Log
 ----------------
 - Version 2.0.3 (pending)
     - Fixed bug that would cause server to crash after running out of files
+    - Added button to print example of all badge types
 
 - Version 2.0.2 (10/10/2017)
     - Split out badge type color bar and age range color bar